Closes INFRA-2082: store all journal fields as strings.
There was an edge case where _CAP_EFFECTIVE was 0 in the first message,
then hex string in subsequent. This set the field mapping to int, and
would fail afterwards. Maybe down the line I should go through all the
journal fields and set them appropriately (int or string), but for now
it seems like a premature optimization with potential for creating more
problems like this one.  There are 2 potential advantages to using int:
smaller storage, and ability to compare. Neither seem like a big deal
right now; sorting can be done on the '@timestamp' field.

This is a breaking change.